+++ src/core/Component.ts
